refactor: Lotto 배열 생성 로직 Lottos로 이동
