+++ src/domain/Lottos.js
@@ -1,8 +1,30 @@
+import Lotto from "./Lotto.js";
+import { UNIT } from "../constant/definition.js";
+import { LOTTO_LENGTH, LOTTO_NUMBER_RANGE } from "../constant/definition.js";
+
 export default class Lottos {
   #lottos;
 
-  constructor(lottos) {
-    this.#lottos = lottos;
+  constructor(purchasePrice) {
+    this.#lottos = this.#createLottos(purchasePrice);
+  }
+
+  #createLottos(purchasePrice) {
+    const lottos = Array.from({ length: purchasePrice / UNIT }).map(
+      () => new Lotto(this.#generateLottoNumbers())
+    );
+    return lottos;
+  }
+
+  #generateLottoNumbers() {
+    const randomNumbers = new Set();
+    while (randomNumbers.size < LOTTO_LENGTH) {
+      const randomNumber = Math.floor(
+        Math.random() * (LOTTO_NUMBER_RANGE.MAX - 1) + 1
+      );
+      randomNumbers.add(randomNumber);
+    }
+    return [...randomNumbers];
   }
 
   getLottosNumber() {
